+++ apps/laboratory/src/utils/ConstantsUtil.ts
@@ -4,6 +4,16 @@ if (!projectId) {
 }
 export const WALLET_URL = process.env['WALLET_URL'] || 'https://react-wallet.walletconnect.com/'
 
+export const CUSTOM_WALLET = 'wc:custom_wallet'
+
+// eslint-disable-next-line init-declarations
+let storedCustomWallet
+if (typeof window !== 'undefined') {
+  storedCustomWallet = localStorage.getItem(CUSTOM_WALLET)
+}
+
+const customWallet = storedCustomWallet ? [JSON.parse(storedCustomWallet)] : []
+
 export const ConstantsUtil = {
   SigningSucceededToastTitle: 'Signing Succeeded',
   SigningFailedToastTitle: 'Signing Failed',
@@ -16,6 +26,7 @@ export const ConstantsUtil = {
     verifyUrl: ''
   },
   CustomWallets: [
+    ...customWallet,
     {
       id: 'react-wallet-v2',
       name: 'react-wallet-v2',