+++ packages/core/src/controllers/AccountController.ts
@@ -35,6 +35,7 @@ export interface AccountControllerState {
   shouldUpdateToAddress?: string
   connectedWalletInfo?: ConnectedWalletInfo
   preferredAccountType?: W3mFrameTypes.AccountType
+  socialWindow?: Window
 }
 
 type StateKey = keyof AccountControllerState
@@ -138,6 +139,12 @@ export const AccountController = {
     }
   },
 
+  setSocialWindow(socialWindow: AccountControllerState['socialWindow']) {
+    if (socialWindow) {
+      state.socialWindow = ref(socialWindow)
+    }
+  },
+
   async fetchTokenBalance() {
     const chainId = NetworkController.state.caipNetwork?.id
 
@@ -174,5 +181,6 @@ export const AccountController = {
     state.connectedWalletInfo = undefined
     state.preferredAccountType = undefined
     state.socialProvider = undefined
+    state.socialWindow = undefined
   }
 }